# Check compilation and linking status only when debug WebGL context is used