@shaojingli 2017-04-25
I did some manual testing. It turns out that benchmark results from `npm run bench` differs significantly from manual benchmark results from a browser. Here are some results for each WebGL function calls during shader compilation and linking:

Node.js: 
gl.shaderSource() took 109310 nanoseconds
gl.compileShader() took 7218032 nanoseconds
gl.getShaderParameter() took 1712 nanoseconds
gl.shaderSource() took 16758 nanoseconds
gl.compileShader() took 536562 nanoseconds
gl.getShaderParameter() took 1419 nanoseconds
gl.attachShader() took 20238 nanoseconds
*gl.linkProgram() took 2458052 nanoseconds*
gl.validateProgram() took 5918 nanoseconds
gl.getProgramParameter() took 804 nanoseconds

Browser (Chrome):
gl.shaderSource() took 0.7799999999988358 milliseconds
gl.compileShader() took 0.22499999999126885 milliseconds
*gl.getShaderParameter() took 17.794999999998254 milliseconds*
gl.shaderSource() took 0.194999999992433 milliseconds
gl.compileShader() took 0.29499999999825377 milliseconds
gl.getShaderParameter() took 2.75 milliseconds
gl.attachShader() took 0.4349999999976717 milliseconds
gl.linkProgram() took 7.169999999998254 milliseconds
*gl.validateProgram() took 94.20500000000175 milliseconds*
gl.getProgramParameter() took 0.41000000000349246 milliseconds