@shaojingli 2017-04-25
@ibgreen @Pessimistress I merged this PR and published a v3.0.2 with this change. Please take a look and benchmark again with our apps.