@ibgreen 2017-04-25
> I did some manual testing. It turns out that benchmark results from npm run bench differs significantly from manual benchmark results from a browser.

That is why there is an `npm run bench-browser`. 

While the call summary is pretty convincing, using the the bench-browser setup, I saw little gain from disabling the shader compilation checks. So I am still not quite convinced removing that check is the right call. The validation check clearly has a big impact on the cached shader scenario.

Without any optimization
```
Program from shader source x 101 ops/sec ±6.79% (49 runs sampled)
Program from cached shaders x 154 ops/sec ±6.33% (51 runs sampled)
```

With only gl.validate disabled
```
Program from shader source x 171 ops/sec ±8.74% (50 runs sampled)
Program from cached shaders x 353 ops/sec ±9.63% (52 runs sampled)
```

With both gl.validate and shader checks disabled
```
Program from shader source x 177 ops/sec ±8.40% (47 runs sampled)
Program from cached shaders x 333 ops/sec ±7.56% (49 runs sampled)
```